Refactor promisify function so we can re-use
