+++ promise.js
@@ -5,7 +5,8 @@ var EventEmitter = require('events').EventEmitter;
 var traverse = require('traverse');
 var Session = require('msgpack5rpc');
 var _ = require('lodash');
-var attachWithCallback = require('.').default;
+var attachWithCallback = require('.');
+var promisify = require('./promisify');
 
 module.exports = function attach(writer, reader) {
   return new Promise(function(resolve, reject) {
@@ -13,35 +14,7 @@ module.exports = function attach(writer, reader) {
       if (err) {
         reject(err);
       } else {
-        //promisify APIs
-        var interfaces = {
-          Nvim: nvim.constructor,
-          Buffer: nvim.Buffer,
-          Window: nvim.Window,
-          Tabpage: nvim.Tabpage,
-        };
-
-        Object.keys(interfaces).forEach(function(key) {
-          var name = key;
-          Object.keys(interfaces[key].prototype).forEach(function(method) {
-            const oldMethod = interfaces[key].prototype[method];
-            interfaces[key].prototype[method] = function() {
-              const args = Array.prototype.slice.call(arguments);
-              const context = this;
-              return new Promise(function(resolve, reject) {
-                args.push(function(err, result) {
-                  if (err) {
-                    reject(err);
-                  } else {
-                    resolve(result);
-                  }
-                });
-                oldMethod.apply(context, args);
-              });
-            };
-          })
-        });
-
+        promisify(nvim);
         resolve(nvim);
       }
     });
